core:Use proper percentage calculation for storage.
Calculate the storage free percentage space correctly, using double
as return value.

Related-to: https://bugzilla.redhat.com/856201

+++ backend/manager/modules/vdsbroker/src/main/java/org/ovirt/engine/core/vdsbroker/irsbroker/IrsBrokerCommand.java
@@ -359,7 +359,7 @@ public abstract class IrsBrokerCommand<P extends IrsBaseVDSCommandParameters> ex
                         || (domainPoolMap.getstatus() != StorageDomainStatus.InActive && data.getstatus() == StorageDomainStatus.Active)) {
                     DbFacade.getInstance().getStorageDomainDynamicDao().update(data.getStorageDynamicData());
                     if (data.getavailable_disk_size() != null && data.getused_disk_size() != null) {
-                        int freePercent = data.getStorageDynamicData().getfreeDiskPercent();
+                        double freePercent = data.getStorageDynamicData().getfreeDiskPercent();
                         int freeDiskInGB = data.getStorageDynamicData().getfreeDiskInGB();
                         AuditLogType type = AuditLogType.UNASSIGNED;
                         boolean spaceThresholdMet =
